refactor: Rename `sdk-js/src/lib/eventDetection`
