+++ sdk-js/src/lib/event-detection/eventTypes/alexaSkill.js
