+++ sdk-js/src/lib/event-detection/eventTypes/apiGateway.js
