+++ sdk-js/src/lib/event-detection/eventTypes/apiGatewayV2.js
