+++ sdk-js/src/lib/event-detection/eventTypes/cloudwatchEvent.js
