+++ sdk-js/src/lib/event-detection/eventTypes/cloudwatchLog.js
