+++ sdk-js/src/lib/event-detection/eventTypes/customAuthorizer.js
