+++ sdk-js/src/lib/event-detection/eventTypes/dynamodb.js
