+++ sdk-js/src/lib/event-detection/eventTypes/firehose.js
