+++ sdk-js/src/lib/event-detection/eventTypes/kinesis.js
