+++ sdk-js/src/lib/event-detection/eventTypes/s3.js
