+++ sdk-js/src/lib/event-detection/eventTypes/ses.js
