+++ sdk-js/src/lib/event-detection/eventTypes/slsIntegrationLambda.js
