+++ sdk-js/src/lib/event-detection/eventTypes/sns.js
