+++ sdk-js/src/lib/event-detection/eventTypes/sqs.js
