+++ sdk-js/src/lib/event-detection/index.js
